Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -6,7 +6,7 @@ int main() {
     int x,y,z;
     scanf("%d %d %d ",&x,&y,&z);
     int sum= x+y+z;
-    float out =sum/3;
+    int out =sum/3.0;
     printf("Average: %.2f",out);
     return 0;
 }
